Breakup sv inspector plugin
