+++ plugins/gff3/src/Gff3TabixAdapter/index.ts
@@ -1,1 +1,15 @@
-export { default as configSchema } from './configSchema'
+import { AdapterType } from '@jbrowse/core/pluggableElementTypes'
+import PluginManager from '@jbrowse/core/PluginManager'
+import configSchema from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'Gff3TabixAdapter',
+        configSchema,
+        getAdapterClass: () =>
+          import('./Gff3TabixAdapter').then(r => r.default),
+      }),
+  )
+}
