+++ plugins/sv-inspector/src/index.ts
@@ -5,60 +5,15 @@ import {
   isAbstractMenuManager,
 } from '@jbrowse/core/util/types'
 import TableChartIcon from '@mui/icons-material/TableChart'
-import SvInspectorViewTypeFactory from './SvInspectorView/SvInspectorViewType'
-import SvInspectorViewModel from './SvInspectorView/models/SvInspectorView'
-
-type SvInspectorView = ReturnType<typeof SvInspectorViewModel>['stateModel']
+import SvInspectorViewF from './SvInspectorView'
+import LaunchSvInspectorViewF from './LaunchSvInspectorView'
 
 export default class SvInspectorViewPlugin extends Plugin {
   name = 'SvInspectorViewPlugin'
 
   install(pluginManager: PluginManager) {
-    pluginManager.addViewType(() =>
-      pluginManager.jbrequire(SvInspectorViewTypeFactory),
-    )
-
-    pluginManager.addToExtensionPoint(
-      'LaunchView-SvInspectorView',
-      // @ts-ignore
-      async ({
-        session,
-        assembly,
-        uri,
-        fileType,
-      }: {
-        session: AbstractSessionModel
-        assembly: string
-        uri: string
-        fileType?: string
-      }) => {
-        // add view, make typescript happy with return type
-        const view = session.addView(
-          'SvInspectorView',
-        ) as unknown as SvInspectorView
-
-        if (!view) {
-          throw new Error('Failed to initialize view')
-        }
-        const exts = uri.split('.')
-        let ext = exts?.pop()?.toUpperCase()
-        if (ext === 'GZ') {
-          ext = exts?.pop()?.toUpperCase()
-        }
-
-        // @ts-ignore
-        view.spreadsheetView.importWizard.setFileType(fileType || ext || '')
-        // @ts-ignore
-        view.spreadsheetView.importWizard.setSelectedAssemblyName(assembly)
-        // @ts-ignore
-        view.spreadsheetView.importWizard.setFileSource({
-          uri,
-          locationType: 'UriLocation',
-        })
-        // @ts-ignore
-        view.spreadsheetView.importWizard.import(assembly)
-      },
-    )
+    SvInspectorViewF(pluginManager)
+    LaunchSvInspectorViewF(pluginManager)
   }
 
   configure(pluginManager: PluginManager) {
